Moved class to namespace where it's used
