+++ rules/downgrade-php72/src/Rector/FunctionLike/AbstractDowngradeReturnDeclarationRector.php
@@ -9,7 +9,6 @@ use PhpParser\Node\Stmt\ClassMethod;
 use PhpParser\Node\Stmt\Function_;
 use Rector\BetterPhpDocParser\PhpDocInfo\PhpDocInfo;
 use Rector\DowngradePhp72\Contract\Rector\DowngradeReturnDeclarationRectorInterface;
-use Rector\DowngradePhp72\Rector\Property\AbstractDowngradeRector;
 use Rector\NodeTypeResolver\Node\AttributeKey;
 
 abstract class AbstractDowngradeReturnDeclarationRector extends AbstractDowngradeRector implements DowngradeReturnDeclarationRectorInterface
