+++ rules/downgrade-php72/src/Rector/FunctionLike/AbstractMaybeAddDocBlockRector.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Rector\DowngradePhp72\Rector\Property;
+namespace Rector\DowngradePhp72\Rector\FunctionLike;
 
 use Rector\Core\Contract\Rector\ConfigurableRectorInterface;
 use Rector\Core\Rector\AbstractRector;
